+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.yffandroid",
+    package = "com.yackfolkfestival",
 )
 
 android_resource(
     name = "res",
-    package = "com.yffandroid",
+    package = "com.yackfolkfestival",
     res = "src/main/res",
 )
