+++ android/app/src/main/java/com/yackfolkfestival/MainActivity.java
@@ -1,4 +1,4 @@
-package com.yffandroid;
+package com.yackfolkfestival;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -13,7 +13,7 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "YFFAndroid";
+        return "YackFolkFestival";
     }
 
     @Override
