+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">YFFAndroid</string>
+    <string name="app_name">Yack Folk Festival</string>
 </resources>
