+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'YFFAndroid'
+rootProject.name = 'YackFolkFestival'
 include ':react-native-screens'
 project(':react-native-screens').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-screens/android')
 include ':react-native-gesture-handler'
