+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "YFFAndroid",
-  "displayName": "YFFAndroid"
+  "name": "YackFolkFestival",
+  "displayName": "Yack Folk Festival"
 }
