+++ constants.js
@@ -14,9 +14,9 @@ module.exports = {
   },
   ENDPOINTS: {
     ALLPERFORMANCES: 'https://raw.githubusercontent.com/RustComet/YFFJSON/master/all_performances_remote.json',
-    ARTISTS: 'https://raw.githubusercontent.com/RustComet/YFFJSON/master/artists_remote.json',
-    FRIPERFORMANCES: 'https://raw.githubusercontent.com/RustComet/YFFJSON/master/fri_performances_remote.json',
-    SATPERFORMANCES: 'https://raw.githubusercontent.com/RustComet/YFFJSON/master/sat_performances_remote.json',
-    SUNPERFORMANCES: 'https://raw.githubusercontent.com/RustComet/YFFJSON/master/sun_performances_remote.json'
+    ARTISTS: 'https://raw.githubusercontent.com/RustComet/YFFJSON/master/staging/artists_remote.json',
+    FRIPERFORMANCES: 'https://raw.githubusercontent.com/RustComet/YFFJSON/master/staging/fri_performances_remote.json',
+    SATPERFORMANCES: 'https://raw.githubusercontent.com/RustComet/YFFJSON/master/staging/sat_performances_remote.json',
+    SUNPERFORMANCES: 'https://raw.githubusercontent.com/RustComet/YFFJSON/master/staging/sun_performances_remote.json'
   }
 };
