+++ ios/YackFolkFestival/Images.xcassets/AppIcon.appiconset/Contents.json
