+++ ios/YackFolkFestivalTests/YackFolkFestivalTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface YFFAndroidTests : XCTestCase
+@interface YackFolkFestivalTests : XCTestCase
 
 @end
 
-@implementation YFFAndroidTests
+@implementation YackFolkFestivalTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
