+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "YFFAndroid",
+  "name": "YackFolkFestival",
   "version": "0.0.1",
   "private": true,
   "scripts": {
